# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/posts": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": " Add new Post",
                "description": " ",
                "operationId": "crate-post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['العنوان مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Post"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "put": {
                "tags": [
                    "posts"
                ],
                "summary": " Update the Post",
                "description": " ",
                "operationId": "update-post",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Example ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['العنوان مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Post"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": " Update the Post",
                "description": " ",
                "operationId": "delete-post",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Example ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error schema",
                "properties": {
                    "title": {
                        "description": "Error Message",
                        "type": "string",
                        "example": "  error meassage "
                    },
                    "desc": {
                        "description": "Error description",
                        "type": "string",
                        "example": "Something went wrong, please try again later."
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "description": "Post schema",
                "properties": {
                    "id": {
                        "description": "The ID of the country",
                        "type": "integer",
                        "example": 59
                    },
                    "title": {
                        "description": "The title of the post",
                        "type": "string",
                        "example": "test title name"
                    },
                    "description": {
                        "description": "The description of the post",
                        "type": "string",
                        "example": "The description of the post "
                    },
                    "created_at": {
                        "description": "The creation timestamp of the service",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-04T21:07:52.000000Z"
                    },
                    "updated_at": {
                        "description": "The update timestamp of the service",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-04T22:29:09.000000Z"
                    },
                    "tags": {
                        "description": "The tags of the post",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
...
            "PostRequest": {
                "title": "Post Request",
                "description": "Post Request  body data",
                "properties": {
                    "title": {
                        "description": "The title of the post",
                        "type": "string",
                        "example": "test title name"
                    },
                    "description": {
                        "description": "The description of the post",
                        "type": "string",
                        "example": "The description of the post "
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "Tag": {
                "title": "Tag",
                "description": "tag schema",
                "properties": {
                    "id": {
                        "description": "The ID of the tag",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the tag",
                        "type": "string",
                        "example": "secure email"
                    },
                    "slug": {
                        "description": "The slug of the tag",
                        "type": "string",
                        "example": "secure-email"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}